feat: add admin table
